S25029301301-1438 抄送我的分页查询接口

+++ onebase-module-bpm/onebase-module-bpm-core/src/main/java/com/cmsr/onebase/module/bpm/core/dal/dataobject/BpmFlowCcRecordDO.java
@@ -36,7 +36,7 @@ public class BpmFlowCcRecordDO extends TenantBaseDO {
      * 已阅 0，否 1，是
      */
     @Column(name = "viewed")
-    private Short viewed;
+    private Integer viewed;
 
     /**
      * 已读时间
